Speed up training: remove .float() encoder cast, increase batch to 64, reduce grad_accum to 2
- BF16 autocast handles DeBERTa FP16 weights safely (no .float() needed)
- batch_size 32→64: A100 40GB has plenty of VRAM
- gradient_accum 4→2: keeps effective batch=128
- Expected ~3-4x speedup

+++ FakeNews-Multimodal-System/src/models/text_transformer.py
@@ -11,9 +11,9 @@ from configs.config import ENHANCED_DIM, HIDDEN_SIZE, MANIP_EMBED_DIM, MODEL_NAM
 class OptimizedMultiTaskModel(nn.Module):
     def __init__(self, model_name: str = MODEL_NAME, dropout_rate: float = 0.1, use_style_in_fake: bool = True) -> None:
         super().__init__()
-        # DeBERTa-v3 stores weights in FP16 internally — convert entire encoder
-        # to FP32 to prevent NaN from mixed-precision gradient flow
-        self.encoder = AutoModel.from_pretrained(model_name).float()
+        # DeBERTa-v3 stores weights in FP16 internally — BF16 autocast handles
+        # precision safely (same exponent range as FP32, no NaN risk)
+        self.encoder = AutoModel.from_pretrained(model_name)
         self.dropout = nn.Dropout(dropout_rate)
         self.use_style_in_fake = use_style_in_fake
